Added csv-write and csv-write-file routines. Updated test and README.
Added routine to convert the CSV data structures generated by the
Text::CSV.parse method back to text strings, and a corresponding
routine to write the CSV strings out to a file. Fixed parse method
to be able to read in fields with white-space on either end. (The
:trim option really wasn't very useful before.) Added support for
using charcters other than default for the quote and separator, both
for parsing and writing. Added support for CSV with empty fields.

Added tests for all new functionality. Valid CSV parsed with
Text::CSV.parse then written back with csv-write should round-trip,
(unless you use on of the mutator options.) Modified all test files
to not use implicit new-lines as it made testing fragile on other
operating systems